feature/dummy

See merge request plugins/cakephp-uppy!1

+++ src/Util/S3Trait.php
@@ -18,20 +18,23 @@ trait S3Trait
      */
     protected function deleteObject(string $path, string $name): bool
     {
-        $s3Client = new \Aws\S3\S3Client(Configure::read('Uppy.S3.config'));
-
-        $exist = $s3Client->doesObjectExist(Configure::read('Uppy.S3.bucket'), $path);
-        if ($exist) {
-            $result = $s3Client->deleteObject([
-                'Bucket' => Configure::read('Uppy.S3.bucket'),
-                'Key' => $path,
-            ]);
-            if ($s3Client->doesObjectExist(Configure::read('Uppy.S3.bucket'), $path)) {
-                return false;
+        if (Configure::read('Uppy.S3.config.connection') === 'dummy') {
+            return true;
+        } else {
+            $s3Client = new \Aws\S3\S3Client(Configure::read('Uppy.S3.config'));
+            $exist = $s3Client->doesObjectExist(Configure::read('Uppy.S3.bucket'), $path);
+            if ($exist) {
+                $result = $s3Client->deleteObject([
+                    'Bucket' => Configure::read('Uppy.S3.bucket'),
+                    'Key' => $path,
+                ]);
+                if ($s3Client->doesObjectExist(Configure::read('Uppy.S3.bucket'), $path)) {
+                    return false;
+                }
             }
-        }
 
-        return true;
+            return true;
+        }
     }
 
     /**
@@ -44,16 +47,18 @@ trait S3Trait
      */
     protected function presignedUrl(string $path, string $name): string
     {
-        $s3Client = new \Aws\S3\S3Client(Configure::read('Uppy.S3.config'));
+        if (Configure::read('Uppy.S3.config.connection') === 'dummy') {
+            return 'https://example.com';
+        } else {
+            $s3Client = new \Aws\S3\S3Client(Configure::read('Uppy.S3.config'));
+            $cmd = $s3Client->getCommand('GetObject', [
+                'Bucket' => Configure::read('Uppy.S3.bucket'),
+                'Key' => $path,
+            ]);
+            $request = $s3Client->createPresignedRequest($cmd, Configure::read('Uppy.S3.contants.lifeTimeGetObject'));
 
-        $cmd = $s3Client->getCommand('GetObject', [
-            'Bucket' => Configure::read('Uppy.S3.bucket'),
-            'Key' => $path,
-        ]);
-
-        $request = $s3Client->createPresignedRequest($cmd, Configure::read('Uppy.S3.contants.lifeTimeGetObject'));
-
-        return (string)$request->getUri();
+            return (string)$request->getUri();
+        }
     }
 
     /**
@@ -66,14 +71,18 @@ trait S3Trait
      */
     protected function createPresignedRequest(string $path, string $contentType): RequestInterface
     {
-        $s3Client = new \Aws\S3\S3Client(Configure::read('Uppy.S3.config'));
-        $command = $s3Client->getCommand('putObject', [
-            'Bucket' => Configure::read('Uppy.S3.bucket'),
-            'Key' => $path,
-            'ContentType' => $contentType,
-            'Body' => '',
-        ]);
+        if (Configure::read('Uppy.S3.config.connection') === 'dummy') {
+            return 'https://example.com';
+        } else {
+            $s3Client = new \Aws\S3\S3Client(Configure::read('Uppy.S3.config'));
+            $command = $s3Client->getCommand('putObject', [
+                'Bucket' => Configure::read('Uppy.S3.bucket'),
+                'Key' => $path,
+                'ContentType' => $contentType,
+                'Body' => '',
+            ]);
 
-        return $s3Client->createPresignedRequest($command, Configure::read('Uppy.S3.contants.lifeTimePutObject'));
+            return $s3Client->createPresignedRequest($command, Configure::read('Uppy.S3.contants.lifeTimePutObject'));
+        }
     }
 }
